chore(import rooms): refactor import mod
